requirements.txt
numpy~=2.2.6
pillow~=12.1.1
osmnx~=2.0.7
geopandas~=1.1.2
earthengine-api~=1.7.15
requests~=2.32.5
mapbox~=0.18.1
